feat: add classic and freestyle mode options
At the moment, these just toggle the ability to lock and delete categories/prompts. We'll have to figure out what happens when you switch between them. Is classic mode always going to have the same four categories or can the user choose from a set number of dropdowns? When switching between modes, should the state of the 'freestyle' inputs persist, or get reset to a default? etc.